refactor(symbolic): split SMT CSE planning
